perf(vm): keep the place instructions out of the dispatch loop
The place model cost `benches/arith` 14.9%, `benches/pure` 10.5% and
`benches/call` 9.3% on programs that execute no place instruction at
all, while the AST column — untouched code, measured in the same session
against commit `aff6550` — stayed within 2% of itself. A capability
nothing uses is not allowed to cost that, so it was attributed rather
than accepted.

Three things carried it, and none of them was the place stack being
touched, because it is not touched.

**`Inst` went from 16 bytes to 24.** `Inst::Call` gained a third
argument count, and three `u32`s beside a `FunctionId` push the widest
variant past the word the rest of them fit in — which is 50% more of
every function's code array, read on every instruction fetch, for a
number that is bounded by the parameters a declaration writes. The
counts are `u16` now, and `call_declared` refuses a declaration with
more parameters than that holds rather than truncating a count into one.
`Inst` is 16 again.

**`Frame` would have gone from 32 bytes to 40.** A frame is copied into
a local of `Vm::execute` and read by every instruction that addresses a
slot, so its width is register pressure in exactly the loop `arith`
spends its run in. `return_pc` is a `u32` now, which is what an
instruction index is everywhere else in the IR — every jump target is
one, and `validate` bounds them all by the code's length — so the frame
is 32 bytes again with the place window in it.

**And `Vm::execute`'s `match` gained seven arms.** That was the largest
of the three, worth about six points of `arith` on its own. They are one
arm now, calling an `#[inline(never)]` function, which is also where
they belong: no benchmark executes a single one of them, so a program
that uses no place pays a jump-table entry and nothing else, and a
program that uses one pays a call it would not have noticed. This is the
ablation study in `docs/VM_ARCHITECTURE.md` read from the other side —
it found `arith` moving several percent for changes that altered nothing
it executed, and this is the same effect with the sign that matters.

Two smaller things came with them. The `Vec<Place>` resize a call makes
and the truncate a return makes are guarded, because a `Vec` whose
element has a destructor is a call with drop glue behind it rather than
the two instructions the scalar stack's is, and almost no callee has a
place window to size. Neither showed up on `arith`, which makes no call
in its loop; together with the frame they are most of what `call` got
back.

What is left is `arith` at +6.7% and everything else at or below +3.9%,
with `arrayget` at −1.6%. Two facts say what that remainder is. The
instruction count and the fuel are identical on both builds —
31,142,877 of each — so it is per-instruction cost and not work. And
the same source change measured through the `cove` binary rather than
through `cove-bench` shows +4.5% rather than +6.7%, which no difference
in work could produce. It is layout, at the edge of the ±6% band the
architecture document already established for `arith`, and nothing
measured here separates it further.

+++ crates/cove-ir/src/lib.rs
@@ -471,11 +471,21 @@ pub enum Inst {
     /// the instruction is written. `validate` reconciles them with the
     /// callee's own `params` and `returns`, which is what makes the
     /// convention an invariant rather than a convention.
+    ///
+    /// The counts are `u16` because this is the widest variant and an enum
+    /// is as wide as its widest one. Three `u32` counts beside a
+    /// [`FunctionId`] made `Inst` 24 bytes where 16 does, which is 50% more
+    /// of every function's code array for a number that cannot use the
+    /// width: an argument count is bounded by the parameters a declaration
+    /// writes, and `crate::lower` refuses a declaration with more of them
+    /// than this holds rather than truncating one. `benches/arith` reads its
+    /// loop out of that array two million times, and it noticed the
+    /// difference.
     Call {
         function: FunctionId,
-        value_argc: u32,
-        scalar_argc: u32,
-        place_argc: u32,
+        value_argc: u16,
+        scalar_argc: u16,
+        place_argc: u16,
         returns_scalar: bool,
     },
     /// Calls a Host operation. `module` and `op` are `Const::Name`.
